requirements.txt
python-dotenv
azure-identity
agent-framework-core
agent-framework-orchestrations
agent-framework-azure-ai
aiohttp
